requirements.txt
# Dependencias principales para la IA
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0

# Machine Learning y NLP
transformers==4.35.0
torch==2.1.0
sentence-transformers==2.2.2
spacy==3.7.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# Base de datos
mysql-connector-python==8.2.0
sqlalchemy==2.0.23

# Utilidades
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-json-logger==2.0.7
aiofiles==23.2.1

# Para Vercel
mangum==0.17.0